pyproject.toml
[tool.poetry]
name = "comics"
version = "0.1.0"
description = ""
authors = ["Aybek18 <kasymaliev09@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
factory-boy = "^3.3.0"
drf-spectacular = "^0.26.5"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pytest-django = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
